pyproject.toml
[tool.poetry]
name = "chiff"
version = "0.1.0"
description = "Chiff command-line client."
authors = ["bas-d <7903735+bas-d@users.noreply.github.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://chiff.dev"
repository = "https://github.com/chiff-app/chiff-cli"
include = [
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.7"
tldextract = "^2.2"
pykeepass = "^3.2"
pynacl = "^1.4"
click = "^7.1"
pillow = "^8.2"
requests = "^2.2"
qrcode = "^6.1"
tabulate = "^0.8.7"
python-daemon = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4"
flake8 = "^3.8.3"
black = {version = "^20.8b1", allow-prereleases = true}
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"

[tool.poetry.scripts]
chiff = "chiff.main:main"
chiffd = "chiff.socket:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

